requirements.txt
ocrmypdf~=16.3
python-sane~=2.9
PyPDF2~=3.0
pydantic~=2.7